refactor: repository ingestor
