+++ plugins/ai-assistant-backend-module-ingestor-azure-devops/src/constants/default-file-types.ts
@@ -1,0 +1,1 @@
+export const defaultFileTypes = ['.md', '.json'];
